Add SafeRunTermination
We have observed that in some cases Geant4 calls internally RunTermination() from the G4MTRunManager. This is would work correctly if Geant4 had control over the event loop, but this is not the case in Gaussino and therefore might result in a situation where the the run has been terminated, but the events are are still created. A new method is added: SafeRunTermination(), which would only be called by Gaussino. An error will be thrown whenever the native RunTermination() is called by G4.